Collision/movement tweaks
Change speed to 24 so that the character moves a cell at a time, and
commented out "setOrigin" as, in this case, it's better that the point
of origin is at the top left.

Added a proof of concept texture to help with the collision, but for
some reason loading the texture in in the map construction causes the
code to lag, even when frame variable is set ridiculously high.