Character selection and projectiles added
We can always change the character select if we decide not to do the
multiple characters.

Projectiles store in a std::vector list (like a C++ arraylist), each
gameLoop it loops through all the Projectile objects in the vector list
thing and calls the updateProjectile method, which moves them by a speed
designated when creating the object. The for loop in the main also
checks for collisions with walls and out of bounds (xPos > 1080 for
example), in which case it erases the object. Which, according to some
reading up, frees the memory for the object too provided the "new"
keyword wasn't used when creating the object.

I put in the out of bounds checks, even though we have a solid wall
around the map, because we might decide to open up a gap in that later.
Who knows.

Anyways, it's coming together, hopefully we can get some enemies in
soon! :D