Added sprite animations for spinning axe
Forgot to mention before, the Projectile class arguments.

isPlayerAttack = whether it's an enemy attack or a player attack. true
for player, false for enemy. Can use it later to calculate damage and
whatnot.

direction: direction that the player is facing, used to decide which way
to send the axe.

userPosition: player vector2f position, used to set starting position of
projectile.

speed: speed or projectile.

missileTexture: I'm loading the texture for the projectile now in the
main, since before it was on the Projectile constructor, meaning it was
reloaded every time a new Projectile object was instantiated, which I
think is a bad idea. :P