Player looks at mouse + throws axe with left click
The player now follows the position of the mouse in the window, and left
click will throw a projectile. Cleaned up the movement code a little bit
in my travels, so that it checks for collisions in the Character class.
Also changed the Projectile/Character.h class so that the LEFT RIGHT UP
DOWN would correspond to the same numbers (ie 3, 1, 0, 2 or whatever)
which made the projectile throwing easier to make accurate.