Map loading added
Sort of, I created a map.setMap(mapChoice) function that changes which
map is being loaded by the object. I also created an array of pointers
to map 2d arrays that I was going to use to make it easier to change the
map, but my Bioshock Infinite download has just finished, so I'm going
to play that for awhile now, haha. If any of you want to try and get
that working, go for it!

Otherwise, if you're fine just calling the setMap function whenever we
want to transition maps, that's fine too. :D