Transitions between maps works now
Also made draw tiles method so that it doesn't draw rectangle shapes for
tileEmpty tiles. Figure this might improve performance a little.

Make a couple gaps for the stores that I'll be working on soonish. I
figure we could easily implement a stat store, where you spend gold to
improve your stats. And items stores for health and stamina potions.

Potions should be easy, just create an int counter on the main, and some
keyboard events for 1 and 2. So if you click these, and healtPotCount >
0, usehealthPot(); healthPotCount--;

Similar for stamina. :)