Experimenting with reducing lag
So, I created an overlay for the black parts of the map on the second
map. Not ideal really. I'd really rather have a
window.clear(sf::Color::Black); and a selective way of resizing the
intRect for the tile2 texture in map, so if anyone wants to try and
tinker with that. Go for it.

Some way to do it dynamically would be neat. Work out the areas of the
map that are set to empty, and create a single sf::rectangleShape to
fill that area with a floor texture.