Lag free maps!
So, I just turned the maps into a single 1080x840 .jpg, and loaded
depending on the curMap (current map).

So now it's completely lag free when running around without enemies. We
should still keep the drawWall method, since commenting that back in
enables us to make changes to the map, and then turn them into another
.jpg in the future. :)