Working potions added and beginning of stores added
So, added useHealthpot and useStampot methods into the Hud class. Also
added some numberofhealthpots/numberofstampots ints into the player
class, along with some getter methods, and removehealthPot/stampot
methods for when one is used.

Started the player off with 1 health pot for testing. Also added a
keyboard even for L in main game loop to deal damage to player, for
testing the use potions methods.

Added some shop screens for the item/vitamin stores, unfinished at the
moment but the look/basic interface elements are there already.

Added an item class, which I'm unsure if I'll use. Pretty sure I'll just
remove it.

Added a use vitamin method into the player class. Takes a char value, h
for health, s for stamina and a for attack. These vitamins permanently
increase the player's stats. So it'll be an easy way for us to put
progression into the game.

However, we should look into putting stat caps in the game. The vitamins
will be quite expensive, but the player could potentially, for example,
gain so much health that the health bar starts going off the side of the
screen.