Music added and changes to character select screen
Kind of lame, but I think it works a little. We can sort something out
nicer if you guys don't like it.

Added that song I linked in chat, we can put more music in or change
where certain music plays later. I just wanted to hear the game with the
music. :)