Comment back in all the "view" related lines to see something cool!
Leaving them out because it messes up the aiming for some reason. Also,
if someone wants to, could you look at the sprite animation? pretty sure
it needs to be on a timer.

The movement of the new sprite looks weirder than the last one, I've
tried to make the movements as exaggerated as possible but at 16*16
they're hard to see.