Sprite animation iterator added and changes to health and stam bars
Made an iterator that counts the number of frames and passes them to the
character::walk method. This allows us to specify how often we want the
animation cycle to change. About every 10 frames looks pretty good.

Also made some changes to where the length of the health and stam bars
are determined in order to make the differences between them more
obvious.

For example, in the previous version a health of 50 was not twice as
long as a stamina of 25.