Character selection works
Made changes to Hud, player, character, projectile and main classes to
reflect your character selection. All handled by a simple int variable
called choice. 0 = warrior, 1 = wizard (should be pretty versatile which
gives room for more characters like thief and cleric if we wanted to).
Changed the colour of the stamina bar to be blue for mana if you are the
wizard, for fun. Feel free to change the blue colour, I'm not entirely
happy with it.