Unlimited Enemies!!!
It's here, the final version of the enemy management system!

Unlimited enemies, unlimited projectiles, unlimited mayhem!!! bring on
the destruction, bring on the pain, THE GAME IS ON

Change log:
- changed how the enemy management works in both the main and enemy
files. Now calculates a set limit of enemies per pass - enemies react
slower when in large numbers but lag issues are handled rather well -
there is a speed drop but that's in the drawing of all the items -
nothing can be done about that!!!