Added enemy sprite
Hey, I get the general idea of how you did the selection character thing
Lance, but it's so different from how I would logic it out that I'm
having trouble getting it perfect. Would you mind editing the setSprite
method to include a third option for the zombie sprite? Might also be a
good idea to move it into the character class. Thanks.