Enemies now have a sprite
Didn't end up changing a whole lot, but the enemies now look like
zombies and the sprite texture is loaded in main.cpp. There are a couple
bugs where only one enemy animates at a time (I think it's due to the
use of &iterations in the walk method but can't figure what the
alternative could be) and they don't animate when the player is holding
down a direction. I've saved a copy of the version from this morning so
it'd be easy enough to revert it if needed.