Bug fix - enemy animations
Moved the iterations reset into the main game loop, because resetting it
in the walk method was causing all sorts of unwanted behaviour (enemies
not animating at all or not animating when you hold down a direction).
Seems to have fixed the problem, let me know if you encounter any more.
Also moved the draw call of the lighting to be after the enemy draw call
so enemies appear out of the shadows all spooky like.