Add dying consequences + music & sound effect
The player can now die (might want to comment out the consequences if
you're play testing. Changed the music around for fun and attempted to
make it so the warrior has more health and the wizard has more mana but
setStamina didn't really work. There is one bug where the input from the
player is registered before the screen has loaded for the game over
state because of the zoomFreeze method, working on a solution...