Action bar added, sounds added, victory messages added
You guys easily add sounds. I've used a buffer for every sf::sound
object (dunno if that's the proper way to do it, but whatever, it
works).

If you want to add a sound say for, a hurt enemy, at the top with the
others put:

sf::SoundBuffer bufferEnemyOuch;
sf::Sound soundEnemyOuch;

And then ctrl-f to find the "loadSounds" method. In there just copy
paste one of the segments down and adjust accordingly! Remember you can
adjust sounds volume with setVolume and pitch with setPitch, if you need
to. :)

Oh and yeah, wherever you want to play the sound just do:
soundEnemyOuch.play();